perf(web): render blog index and posts as server components
/blog and /blog/[id] were client components importing every post body, so ~250kB of markdown plus react-markdown/remark/marked shipped as JS. Move rendering to the server: /blog/[id] is now statically generated (generateStaticParams) with the article rendered server-side, and /blog passes only card metadata to a small client island. Interactive bits (scroll-spy TOC, smooth scroll) are isolated client islands. First-load JS: /blog/[id] 1584->505kB, /blog 883->529kB.
